+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Support multiple webhook endpoints per project with different configuration
 - Per outgoing webhook event configuration
+- Configuration of secret via ENV variable
 
 ### Changed
 - Skip update events without changelog